handled javaa execution errors

+++ rce-server/server/api/services/code.service.js
@@ -7,11 +7,12 @@ const ROOT_DIR = `${process.cwd()}`;
 const SOURCE_DIR = path.join(ROOT_DIR, 'executor');
 const TARGET_DIR = `/app/codes`;
 const IMAGE_NAME = 'executor:1.0';
-const VOL_NAME = `my_vol`;
-//const VOL_NAME = SOURCE_DIR;
+//const VOL_NAME = `my_vol`;
+const VOL_NAME = SOURCE_DIR;
 
 class CodeService {
   async execute(code, input, lang, id) {
+    //console.log('code', code);
     try {
       !input ? (input = '') : null;
 
@@ -51,6 +52,7 @@ class CodeService {
       );
 
       if (OUTPUT) {
+        console.log('output', OUTPUT.toString());
         return OUTPUT.toString();
       }
     } catch (error) {
@@ -65,7 +67,7 @@ class CodeService {
         fileName += '.js';
         break;
       }
-      case 'c++': {
+      case 'cpp': {
         fileName += '.cpp';
         break;
       }
@@ -106,7 +108,7 @@ class CodeService {
         command = `cd ${TARGET_DIR} && node ${file} < ${input}`;
         break;
       }
-      case 'c++': {
+      case 'cpp': {
         command = `cd ${TARGET_DIR} && g++ -o ${id} ${file} && ./${id} < ${input}`;
         break;
       }
@@ -165,15 +167,17 @@ class CodeService {
         if (err) throw { message: err };
       });
     }
-    if (lang == 'c++' || lang == 'c') {
-      fs.unlinkSync(path.join(SOURCE_DIR, id), err => {
-        if (err) throw err;
-      });
+    if (lang == 'cpp' || lang == 'c') {
+      if (fs.existsSync(path.join(SOURCE_DIR, id)))
+        fs.unlinkSync(path.join(SOURCE_DIR, id), err => {
+          if (err) throw err;
+        });
     }
     if (lang == 'java') {
-      fs.unlinkSync(path.join(SOURCE_DIR, 'Input.class'), err => {
-        if (err) throw err;
-      });
+      if (fs.existsSync(path.join(SOURCE_DIR, 'Input.class')))
+        fs.unlinkSync(path.join(SOURCE_DIR, 'Input.class'), err => {
+          if (err) throw err;
+        });
     }
   }
